requirements.txt
Django==1.8
dj-database-url==0.3.0
psycopg2==2.6
django-flat-theme==0.9.3
flake8==2.4.0
coverage==3.7.1
coveralls==0.5
django-markupfield==1.3.0
Markdown==2.6.1